Доделал меню покупки, пофиксил пролёт объектов в дэше,
Мелкие фиксы

+++ lib/components/DBHandler.dart
@@ -165,6 +165,11 @@ class DbHandler
   Future fillGameObjects() async
   {
     for(final wrld in fullMaps()){
+      final res = await _database?.rawQuery('SELECT COUNT(*) as count FROM ${wrld.nameForGame}');
+      if(res![0]['count'] as int != 0){
+        print('full objects');
+        continue;
+      }
       await _database?.execute('DELETE FROM ${wrld.nameForGame}');
       try {
         var data = await rootBundle.loadString(
